% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageb}
\alias{ageb}
\title{Polygons of AGEB from Coatepec}
\format{
A sf object with 18 rows and 2 variables:
\describe{
\item{CVEGEO}{the ids.}
\item{geometry}{simple feature.}
...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
ageb
}
\description{
A spatial dataset containing of polygon data of agebs from Coatepec
}
\details{
This datasets were published in 2025.
}
\keyword{datasets}
