% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageb_ver}
\alias{ageb_ver}
\title{Agebs veracruz}
\format{
A sf object with 3346 rows and 6 variables:
\describe{
\item{CVEGEO}{the string key. The key included the CVE_ENT, CVE_MUN, CVE_LOC and CVE_AGEB}
\item{CVE_ENT}{is the string key for each state.}
\item{CVE_MUN}{is the string key to the Municipality.}
\item{CVE_LOC}{is string the key of locality.}
\item{CVE_AGEB}{is the string key of AGEB.}
\item{geometry}{simple feature.}
}
}
\usage{
ageb_ver
}
\description{
A spatial dataset containing polygon data of AGEBs from Veracruz.
}
\details{
This datasets were published in 2025.
}
\keyword{datasets}
