% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_ageb.R
\name{extrae_ageb}
\alias{extrae_ageb}
\title{Extrae agebs}
\usage{
extrae_ageb(data_ageb, data_loc, localidad)
}
\arguments{
\item{data_ageb}{es la base de ageb del estado}

\item{data_loc}{es la base de localidades del estado}

\item{localidad}{es la localidad de interes}
}
\value{
un objeto sf
}
\description{
Extrae agebs
}
\examples{
1+1
}
