% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_ver}
\alias{loc_ver}
\title{localidades Veracruz}
\format{
A sf object with 5812 rows and 7 variables:
\describe{
\item{CVEGEO}{the string key. The key included the CVE_ENT, CVE_MUN, CVE_LOC and CVE_AGEB}
\item{CVE_ENT}{is the string key for each state.}
\item{CVE_MUN}{is the string key to the Municipality.}
\item{CVE_LOC}{is string the key of locality.}
\item{CVE_AGEB}{is the string key of AGEB.}
\item{Ambito}{is a binary variable with two values, Urbana and Rural. The dataset is inly AGEB Urban}
\item{geometry}{simple feature.}
}
}
\usage{
loc_ver
}
\description{
A spatial dataset containing polygon data of localities from Veracruz.
}
\details{
This datasets were published in 2025.
}
\keyword{datasets}
